changed public interface and export structure
