+++ src/crack/worker_threads.rs
@@ -8,8 +8,7 @@ use std::thread;
 use std::thread::JoinHandle;
 
 use crate::crack::indices::{indices_create, indices_increment_by, indices_to_string};
-use crate::crack::parameter::InternalCrackParameter;
-use crate::CrackTarget;
+use crate::{CrackTarget, InternalCrackParameter};
 
 /// Spawns all worker threads.
 pub(crate) fn spawn_worker_threads<T: CrackTarget>(
