+++ src/lib.rs
@@ -4,12 +4,16 @@
 //! new things about Rust.
 //!
 //! # Minimal example
-//! ```ignore
+//! ```rust
 //! use libbruteforce::{BasicCrackParameter, CrackParameter, TargetHashInput};
 //! use libbruteforce::hash_fncs::sha256_hashing;
+//! use libbruteforce::symbols;
 //!
 //! // sha256("a+c")
 //! let sha256_hash = "3d7edde33628331676b39e19a3f2bdb3c583960ad8d865351a32e2ace7d8e02d";
+//! let max_len = 3;
+//! let min_len = 0;
+//! let alphabet = symbols::Builder::new().with_digits().build();
 //!
 //! let res = CrackParameter::new(
 //!         BasicCrackParameter::new(alphabet, max_len, min_len, false),
@@ -35,17 +39,23 @@
 #![deny(rustdoc::all)]
 #![allow(rustdoc::missing_doc_code_examples)]
 
-pub use crack::crack;
-pub use crack::parameter::{BasicCrackParameter, CrackParameter};
-use std::fmt::{Debug, Formatter};
+use crate::crack::worker_threads::spawn_worker_threads;
+use std::fmt::Debug;
+use std::sync::atomic::AtomicBool;
+use std::sync::Arc;
+use std::time::Instant;
+
+pub use crack::CrackResult;
+pub use parameter::*;
 
 mod crack;
 #[cfg(test)]
 mod testutil;
 
 // Public API
-pub mod symbols;
 pub mod hash_fncs;
+mod parameter;
+pub mod symbols;
 
 /// Common trait for crack targets (hashes or plain text to crack). This is the super-type
 /// which enables the usage of multiple hashing algorithms. An example that
@@ -59,106 +69,129 @@ pub trait CrackTarget: 'static + Eq + Send + Sync + Debug {}
 // automatically impl the trait for all types that fulfill the condition/required traits
 impl<T> CrackTarget for T where T: 'static + Eq + Send + Sync + Debug {}
 
-/// Helper type to create instances of [`TargetHashAndHashFunction`].
-#[derive(Debug)]
-pub enum TargetHashInput<'a> {
-    /// The provided input is already a valid hash but as (hex) string representation.
-    HashAsStr(&'a str),
-    /// The provided input is plain text and needs to be hashed by the constructor.
-    /// This is useful for tests, examples, and debugging. For real applications you
-    /// may want to use [`Self::HashAsStr`].
-    Plaintext(&'a str),
+/// This function starts a multi-threaded brute force attack on a given target string. It supports
+/// any alphabet you would like to use. You must provide a hashing function. There is a pre-build
+/// set of transformation functions available, such as [`hash_fncs::no_hashing`] or
+/// [`hash_fncs::sha256_hashing`]. You can also provide your own hashing strategy.
+///
+/// This library is really "dumb". It checks each possible value and doesn't use any probabilities
+/// for more or less probable passwords.
+///
+/// # Parameters
+/// * `param` - See [`CrackParameter`]
+///
+/// # Return
+/// Returns a [`CrackResult`].
+pub fn crack<T: CrackTarget>(param: CrackParameter<T>) -> CrackResult {
+    let param = InternalCrackParameter::from(param);
+    let param = Arc::from(param);
+
+    // shared atomic bool so that all threads can look if one already found a solution
+    // so they can stop their work. This only gets checked at every millionth iteration
+    // for better performance.
+    let done = Arc::from(AtomicBool::from(false));
+    let instant = Instant::now();
+    let handles = spawn_worker_threads(param.clone(), done);
+
+    // wait for all threads
+    let solution = handles
+        .into_iter()
+        .map(|h| h.join().unwrap()) // result of the Option<String> from the threads
+        .flatten()
+        .last(); // extract from the collection
+
+    let seconds = instant.elapsed().as_secs_f64();
+
+    let param =
+        Arc::try_unwrap(param).unwrap_or_else(|_| panic!("There should only be one reference!"));
+    if let Some(solution) = solution {
+        CrackResult::new_success(param, seconds, solution)
+    } else {
+        CrackResult::new_failure(param, seconds)
+    }
 }
 
-/// Abstraction over a hashing algorithm and the target hash that needs to be cracked.
-/// `T` is of type [`CrackTarget`]. This generic struct exists so that hashes of type
-/// [`CrackTarget`] can be checked independent of the hashing algorithm. This is
-/// more efficient than transforming every hash to a string and compare the hash
-/// string representations afterwards.
-pub struct TargetHashAndHashFunction<T: CrackTarget> {
-    /// The target hash we want to crack.
-    target_hash: T,
-    /// Function that calculates the hash of type `T` of the given input plain text.
-    hash_fn: fn(input: &str) -> T,
-    /// Function that transforms a `T` in string representation to a real `T`.
-    /// For example, this transforms a `sha256` string representation to the runtime
-    /// type the hashing library uses.
-    hash_str_repr_to_hash_type_fn: fn(hash_as_string: &str) -> T,
-    /// Function that transform the hash type to a string representation. Usually, this
-    /// will return a hex string that represents the hash.
-    hash_type_to_str_repr_fn: fn(hash: &T) -> String
-}
+#[cfg(test)]
+mod tests {
+    use crate::hash_fncs::no_hashing;
+    use crate::testutil::{
+        create_test_crack_params_full_alphabet, create_test_crack_params_full_alphabet_sha256,
+        create_test_crack_params_full_alphabet_sha256_fair,
+    };
+    use crate::{BasicCrackParameter, CrackParameter, TargetHashInput};
 
-impl<T: CrackTarget> TargetHashAndHashFunction<T> {
-    /// Constructor that takes a hashing function and a target hash.
-    ///
-    /// # Parameters
-    /// * `target_hash` String representation of the target hash we want to crack.
-    ///                 This is usually the hex string representation of a sha256 hash or so.
-    /// * `hash_fn` Transforms a plain input password/guess of type `str` to the target hash.
-    ///                  This is the hashing function.
-    /// * `hash_str_repr_to_hash_type_fn` Function that can take the argument `target_hash`
-    ///                                   and transform it to the target hashing type. This
-    ///                                   usually transforms the hex string that represents the
-    ///                                   hash to bytes in memory.
-    /// * `hash_type_to_str_repr_fn` Function that transform the hash type to a string representation.
-    ///                             Usually, this will return a hex string that represents the hash.
-    pub fn new(
-        target_hash: TargetHashInput,
-        hash_fn: fn(&str) -> T,
-        hash_str_repr_to_hash_type_fn: fn(hash_as_string: &str) -> T,
-        hash_type_to_str_repr_fn: fn(hash: &T) -> String
-    ) -> Self {
-        let target_hash = match target_hash {
-            TargetHashInput::HashAsStr(hash_str) => hash_str_repr_to_hash_type_fn(hash_str),
-            TargetHashInput::Plaintext(input) => hash_fn(input)
-        };
-        Self {
-            target_hash,
-            hash_fn,
-            hash_str_repr_to_hash_type_fn,
-            hash_type_to_str_repr_fn
+    use super::*;
+
+    #[test]
+    #[should_panic]
+    fn test_crack_should_panic_1() {
+        let cp = CrackParameter::new(
+            BasicCrackParameter::new(vec!['a'].into_boxed_slice(), 4, 5, false),
+            no_hashing(TargetHashInput::Plaintext("a+c")),
+        );
+        // expect panic; min > max
+        let _res = crack(cp);
+    }
+
+    #[test]
+
+    fn test_crack_dont_find_bc_of_min_length() {
+        let cp = CrackParameter::new(
+            BasicCrackParameter::new(vec!['a'].into_boxed_slice(), 4, 3, false),
+            no_hashing(TargetHashInput::Plaintext("a+c")),
+        );
+        // expect panic; min > max
+        let res = crack(cp);
+        assert!(
+            !res.is_success(),
+            "should not find result, because of min length!"
+        );
+    }
+
+    #[test]
+    fn test_crack_identity() {
+        let input = String::from("a+c");
+        let target = input.clone(); // identity hashing
+        let cp = create_test_crack_params_full_alphabet(&target);
+        let res = crack(cp);
+        assert!(res.is_success(), "a solution must be found!");
+        assert!(
+            input.eq(res.solution().as_ref().unwrap()),
+            "target and cracked result must equal!"
+        );
+    }
+
+    /// Warning: can run a few seconds (15-30)
+    #[test]
+    fn test_crack_sha256_fair_and_unfair_mode() {
+        let input = "a+c";
+        let target =
+            String::from("3d7edde33628331676b39e19a3f2bdb3c583960ad8d865351a32e2ace7d8e02d");
+
+        let cp = create_test_crack_params_full_alphabet_sha256(&target);
+        let cp_fair = create_test_crack_params_full_alphabet_sha256_fair(&target);
+
+        let res = crack(cp);
+        assert!(res.is_success(), "A solution MUST be found!");
+        assert!(
+            input.eq(res.solution().as_ref().unwrap()),
+            "The cracked value is wrong! It's"
+        );
+
+        if num_cpus::get() > 1 {
+            assert!(cp_fair.fair_mode(), "Fair mode must be activated"); // check if really multiple threads are used
+            assert!(res.thread_count() > 1, "multiple threads must be used"); // check if really multiple threads are used
+
+            let res_fair = crack(cp_fair);
+            assert!(res_fair.is_success(), "A solution MUST be found!");
+            assert!(
+                input.eq(res.solution().as_ref().unwrap()),
+                "The cracked value is wrong!"
+            );
+            assert!(
+                res.thread_count() > res_fair.thread_count(),
+                "fair mode must use less treads!"
+            );
         }
-    }
-}
-
-impl<T: CrackTarget> Debug for TargetHashAndHashFunction<T> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("TargetHashAndHashFunction")
-            .field("target_hash", &self.target_hash)
-            .field("hash_fn", &"<func impl>")
-            .field("target_str_repr_to_hash_type_fn", &"<func impl>")
-            .field("hash_type_to_str_repr_fn", &"<func impl>")
-            .finish()
-    }
-}
-
-impl<T: CrackTarget> TargetHashAndHashFunction<T> {
-
-    /// Transforms the (hex) string representation into the type
-    /// the hash implementation uses to represent hashes.
-    pub fn hash_str_repr_to_hash_type(&self, hash_as_string: &str) -> T {
-        (self.hash_str_repr_to_hash_type_fn)(hash_as_string)
-    }
-
-    /// Hashes a value.
-    pub fn hash(&self, input: &str) -> T {
-        (self.hash_fn)(input)
-    }
-
-    /// Returns the target hash that we want to crack.
-    pub fn target_hash(&self) -> &T {
-        &self.target_hash
-    }
-
-    /// Returns a (hex) string representation of the hash.
-    pub fn hash_type_to_str_repr(&self, hash: &T) -> String {
-        (self.hash_type_to_str_repr_fn)(hash)
-    }
-
-    /// Hashes the input value and returns if it equals the target hash.
-    /// If so, the hash got cracked.
-    pub fn hash_matches(&self, input: &str) -> bool {
-        (self.hash_fn)(input) == self.target_hash
     }
 }
