+++ src/symbols/mod.rs
@@ -39,7 +39,7 @@ pub static ALL_OTHER_SPECIAL_CHARS: [char; 23] = [
 
 /// Calculates the amount of possible permutations if n symbols are given and m slots are available.
 /// This solutions counts in that the value can be zero-length, one-length and so on.
-pub fn combinations_count(alphabet: &[char], max_length: u32, min_length: u32) -> usize {
+pub fn combination_count(alphabet: &[char], max_length: u32, min_length: u32) -> usize {
     if min_length > max_length {
         panic!("max_length must be >= min_length")
     }
@@ -62,38 +62,38 @@ mod tests {
         let alphabet4: Box<[char]> = Box::from([]);
         let alphabet5: Box<[char]> = Box::from(DIGITS);
         assert_eq!(
-            combinations_count(&alphabet1, 3, 0),
+            combination_count(&alphabet1, 3, 0),
             4,
             "1 symbol and a maximum length of 3"
         );
         assert_eq!(
-            combinations_count(&alphabet1, 3, 3),
+            combination_count(&alphabet1, 3, 3),
             1,
             "1 symbol, min=2, max=3 => 2 combination"
         );
         assert_eq!(
-            combinations_count(&alphabet1, 3, 3),
+            combination_count(&alphabet1, 3, 3),
             1,
             "1 symbol, min=3, max=3 => 1 combination"
         );
         assert_eq!(
-            combinations_count(&alphabet2, 1, 0),
+            combination_count(&alphabet2, 1, 0),
             4,
             "3 symbols and a maximum length of 1"
         );
         assert_eq!(
-            combinations_count(&alphabet2, 2, 1),
+            combination_count(&alphabet2, 2, 1),
             12,
             "3 symbols, min=1, max=2 => 9 combination"
         );
         assert_eq!(
-            combinations_count(&alphabet3, 3, 0),
+            combination_count(&alphabet3, 3, 0),
             15,
             "3 symbols and a maximum length of 3"
         );
-        assert_eq!(combinations_count(&alphabet4, 0, 0), 1, "0 symbols");
-        assert_eq!(combinations_count(&alphabet4, 0, 0), 1, "0 symbols");
-        assert_eq!(combinations_count(&alphabet5, 4, 4), 10_000);
+        assert_eq!(combination_count(&alphabet4, 0, 0), 1, "0 symbols");
+        assert_eq!(combination_count(&alphabet4, 0, 0), 1, "0 symbols");
+        assert_eq!(combination_count(&alphabet5, 4, 4), 10_000);
     }
 
     #[test]
@@ -101,7 +101,7 @@ mod tests {
     fn test_combinations_count_panic() {
         let alphabet: Box<[char]> = Box::from(['a']);
         assert_eq!(
-            combinations_count(&alphabet, 0, 1),
+            combination_count(&alphabet, 0, 1),
             0,
             "min length must be <= max length"
         );
